finhub_newsfeed_websockets/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.2
asyncio-nats-client==0.11.5
asyncpg==0.27.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
dataclasses-json==0.5.7
dnspython==2.3.0
docx2txt==0.8
exceptiongroup==1.1.2
fastapi==0.99.1
fredapi==0.5.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
langchain==0.0.192
langchainplus-sdk==0.0.4
loguru==0.7.0
lxml==4.9.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
Pillow==10.0.0
pinecone-client==2.2.2
protobuf==3.20.3
pydantic==1.10.8
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
websocket-client==1.6.2
websockets==11.0.3
XlsxWriter==3.1.2
yarl==1.9.2
